api-backend/pyproject.toml
[project]
name = "api-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "asyncpg>=0.31.0",
    "better-auth>=0.0.1b11",
    "better-exceptions>=0.3.3",
    "fastapi>=0.124.0",
    "openai>=2.9.0",
    "openai-agents>=0.6.2",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.16.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "sentence-transformers>=5.1.2",
    "sqlalchemy>=2.0.44",
    "torch>=2.9.1",
    "transformers>=4.57.3",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "faker>=38.2.0",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "testcontainers>=4.13.3",
]
